96 make meter api able to save data to the db (#97)
* GC Suppress

* Swap to double and implement snapshot service

* add post put and patch methods

* spacing

* improve gfunction

+++ meter-api/meter-api/Datatypes/MeterSnapshot.cs
@@ -4,8 +4,8 @@
     { 
         public required string MeterId { get; set; }
         public required string DisplayName { get; set; }
-        public required double CurrentUsage { get; set; }
-        public required double TotalUsage { get; set; }
+        public required decimal CurrentUsage { get; set; }
+        public required decimal TotalUsage { get; set; }
         public required decimal TotalCost { get; set; }
     }
 }
